+++ pim/src/PcmtCoreBundle/Connector/Job/JobParameters/ConstraintCollectionProvider/ReferenceDataXmlImportProvider.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * Copyright (c) 2020, VillageReach
+ * Licensed under the Non-Profit Open Software License version 3.0.
+ * SPDX-License-Identifier: NPOSL-3.0
+ */
+
+declare(strict_types=1);
+
+namespace PcmtCoreBundle\Connector\Job\JobParameters\ConstraintCollectionProvider;
+
+use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
+use PcmtCoreBundle\Connector\Job\JobParameters\SupportedJobsTrait;
+use Symfony\Component\Validator\Constraints\Collection;
+
+class ReferenceDataXmlImportProvider implements ConstraintCollectionProviderInterface
+{
+    use SupportedJobsTrait;
+
+    public function __construct(
+        array $supportedJobNames
+    ) {
+        $this->supportedJobNames = $supportedJobNames;
+    }
+
+    public function getConstraintCollection(): Collection
+    {
+        return new Collection([
+            'fields' => [
+                'filePath' => [],
+            ],
+        ]);
+    }
+}